+++ src/extension.js
@@ -1,5 +1,6 @@
 import { registerSaveTracker } from "./save-tracker.js";
 import { registerSaveCommands } from "./save.js";
+import { registerStatusbarItem } from "./status.js";
 
 let activated = false;
 
@@ -10,6 +11,8 @@ export function activate(context) {
 	if (activated) return;
 
 	activated = true;
+
+	registerStatusbarItem(context);
 
 	registerSaveTracker(context);
 	registerSaveCommands(context);