+++ src/save.js
@@ -1,7 +1,7 @@
 import vscode from 'vscode';
 
-import { getSaves, clearSaves } from './save-tracker.js';
-import { formatDate } from './time.js';
+import { getSavesFromState, clearSaves } from './save-tracker.js';
+import { formatDate, formatHour, formatNTimes } from './format.js';
 
 /**
  * @param {vscode.ExtensionContext} context
@@ -16,7 +16,7 @@ export function registerSaveCommands(context) {
         }).then(document => {
             vscode.window.showTextDocument(document);
         });
-	}, null, context.subscriptions);
+    }, null, context.subscriptions);
 
     vscode.commands.registerCommand('ctrl-s.resetStatistics', () => {
         vscode.window.showInformationMessage('Are you sure you want to reset your statistics?', 'Yes', 'No').then(answer => {
@@ -24,11 +24,11 @@ export function registerSaveCommands(context) {
                 clearSaves();
             }
         });
-	}, null, context.subscriptions);
+    }, null, context.subscriptions);
 }
 
 function buildStatisticsString() {
-    const saves = getSaves();
+    const saves = getSavesFromState();
 
     let totalSaves = 0,
         days = {},
@@ -68,13 +68,9 @@ function buildStatisticsString() {
     }
 
     const earliestSave = new Date(Math.min(...Object.keys(saves).map(timestamp => timestamp * 1000))),
-        earliestSaveDate = formatDate(earliestSave);
-
-    const averageSavesPerDay = Math.round(totalSaves / Object.keys(days).length),
+        averageSavesPerDay = Math.round(totalSaves / Object.keys(days).length),
         mostCommonHour = Object.keys(hours).reduce((a, b) => hours[a] > hours[b] ? a : b),
         mostCommonLanguage = Object.keys(languages).reduce((a, b) => languages[a] > languages[b] ? a : b);
 
-    const mostCommonHourDate = mostCommonHour === 0 ? 'midnight' : mostCommonHour === 12 ? 'noon' : `${mostCommonHour > 12 ? mostCommonHour - 12 : mostCommonHour} ${mostCommonHour > 12 ? 'PM' : 'AM'}`;
-
-    return `You have saved ${totalSaves} times since ${earliestSaveDate}, averaging about ${averageSavesPerDay} saves per day. You most commonly save ${mostCommonLanguage} files at ${mostCommonHourDate}.`;
+    return `You have saved ${formatNTimes(totalSaves)} since ${formatDate(earliestSave)}, averaging about ${formatNTimes(averageSavesPerDay)} a day. You most commonly save ${mostCommonLanguage} files at ${formatHour(mostCommonHour)}.`;
 }